+++ source/digioz.Forum/digioz.Forum/Models/ForumSearchResult.cs
@@ -9,7 +9,7 @@ public partial class ForumSearchResult
     [Key]
     public string SearchKey { get; set; } = null!;
 
-    public int SearchTime { get; set; }
+    public long SearchTime { get; set; }
 
     public string SearchKeywords { get; set; } = null!;
