+++ source/digioz.Forum/digioz.Forum/Models/ForumTrack.cs
@@ -9,5 +9,5 @@ public partial class ForumTrack
 
     public long ForumId { get; set; }
 
-    public int MarkTime { get; set; }
+    public long MarkTime { get; set; }
 }
